@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "engines": {
-    "node": "12.x"
+    "node": "20.x"
   },
   "dependencies": {
     "@babel/core": "^7.15.8",
@@ -22,10 +22,12 @@
     "formik": "^2.2.9",
     "i18next-browser-languagedetector": "^6.1.2",
     "i18next-http-backend": "^1.2.6",
+    "react": "^17.0.2",
     "react-i18next": "^11.11.2",
     "react-icons": "^4.3.1",
     "react-redux": "^7.2.4",
-    "react-router-dom": "^5.3.4",
+    "react-router": "^5.3.0",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "^5.0.1",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.3.0",
