@@ -1,14 +1,12 @@
-import { useEffect, useState } from "react";
-import * as yup from "yup";
-import { Formik } from "formik";
-import Button from "@material-ui/core/Button";
-import TextField from "@material-ui/core/TextField";
+import * as yup from 'yup';
+import { Formik } from 'formik';
+import Button from '@material-ui/core/Button';
+import TextField from '@material-ui/core/TextField';
 // import CheckBoxOutlineBlankIcon from "@material-ui/icons/CheckBoxOutlineBlank";
 // import CheckBoxIcon from "@material-ui/icons/CheckBox";
-import Autocomplete from "@material-ui/lab/Autocomplete";
-import styles from "./CarrierPage.module.css";
-import { uploadTransfer } from "../../../trip_search/data/api/data-service";
-import { useHistory } from "react-router-dom";
+import Autocomplete from '@material-ui/lab/Autocomplete';
+import styles from './CarrierPage.module.css';
+import { useHistory } from 'react-router-dom';
 import {
   Checkbox,
   Container,
@@ -24,51 +22,51 @@ import {
   Paper,
   Select,
   Tooltip,
-} from "@material-ui/core";
-import cities_json from "../../../trip_search/domain/entites/utils/jsons/cities.json";
-import i18n from "../../../trip_search/domain/entites/utils/language/i18n";
-import { useSelector } from "react-redux";
-import { currencies } from "../currenciesSelector/currencies";
-import axios from "axios";
-import "yup-phone-lite";
-import { timeZones } from "./timezones/timezones";
-import { getLoading } from "../../../trip_search/presentation/redux/reducers/selectors";
-import { useDispatch } from "react-redux";
-import { LoadingButton } from "@mui/lab";
-import { Alert } from "@mui/material";
-import { loadingUploadTransferAction } from "../../../trip_search/presentation/redux/reducers/actions/loading-actions";
-import {useStyles} from "../../../../general/MUI/useStyles";
+} from '@material-ui/core';
+import i18n from '../../../trip_search/domain/entites/utils/language/i18n';
+import { currencies } from '../currenciesSelector/currencies';
+import 'yup-phone-lite';
+import { LoadingButton } from '@mui/lab';
+import { Alert } from '@mui/material';
+import { useStyles } from '../../../../general/MUI/useStyles';
+import useCarrier from '../hooks/useCarrier';
 
 const schema = yup.object().shape({
-  from: yup.string().required("from.Required"),
-  to: yup.string().required("to.Required"),
-  date: yup.date().required("date.Required"), //!!! date, departureTime, and duration are conditionally reassigned later;
-  departureTime: yup.string().required("departureTime.Required"),
+  from: yup.string().required('from.Required'),
+  to: yup.string().required('to.Required'),
+  date: yup.date().required('date.Required'), //!!! date, departureTime, and duration are conditionally reassigned later;
+  departureTime: yup.string().required('departureTime.Required'),
   // duration: yup.string().required("duration.Required"),
   places: yup
     .number()
-    .min(1, "Available places must be more or equal to 1")
-    .max(8, "Available places must be less or equal to 8")
-    .required("places.Required"),
-  phoneNumber: yup.string().required("phoneNumber.Required").phone(undefined, "phoneNumber.isNotValid"),
-  price: yup.number().positive("price.OnlyPositive"),
+    .min(1, 'Available places must be more or equal to 1')
+    .max(8, 'Available places must be less or equal to 8')
+    .required('places.Required'),
+  phoneNumber: yup
+    .string()
+    .required('phoneNumber.Required')
+    .phone(undefined, 'phoneNumber.isNotValid'),
+  price: yup.number().positive('price.OnlyPositive'),
 });
 
 export default function CarrierPage() {
-  const dispatch = useDispatch();
-  const cur = useSelector((state) => state.app.currency);
-  const lang = useSelector((state) => state.app.lang);
-  const loading = useSelector(getLoading).isLoadingNewTransfer;
+  const {
+    loading,
+    rideCurrency,
+    setRideCurrency,
+    cities,
+    message,
+    open,
+    setOpen,
+    latitude,
+    longitude,
+    submitForm,
+    getDefaultCity,
+    userTimeZone,
+    durations
+  } = useCarrier();
+
   const classes = useStyles();
-  const [rideCurrency, setRideCurrency] = useState(cur);
-  // const [messenger, setMessenger] = useState();
-  // const [nearestCity, setNearestCity] = useState();
-  const [latitude, setLatitude] = useState();
-  const [longitude, setLongitude] = useState();
-  const [userTimeZone, setUserTimeZone] = useState(() => {
-    const timeZone = timeZones.find((tz) => tz.shift === "" + new Date().getTimezoneOffset() / -60);
-    return timeZone || timeZones[0];
-  });
 
   // console.log("cur: ", cur);
   // console.log("rideCurrency: ", rideCurrency);
@@ -77,48 +75,6 @@ export default function CarrierPage() {
   // console.log(lang);
   //
   // console.log("user time zone", userTimeZone);
-
-  const cities =
-    lang === "ru"
-      ? [
-          ...cities_json
-            .reduce((acc, val) => {
-              acc.push({ id: val.ID, title: val["name_ru"] });
-              return acc;
-            }, [])
-            .sort((a, b) => (a.title < b.title ? -1 : 1)),
-          ...cities_json
-            .reduce((acc, val) => {
-              acc.push({ id: val.ID, title: val.name });
-              return acc;
-            }, [])
-            .sort((a, b) => (a.title < b.title ? -1 : 1)),
-        ]
-      : cities_json
-          .reduce((acc, val) => {
-            acc.push({ id: val.ID, title: val.name });
-            acc.push({ id: val.ID, title: val["name_ru"] });
-            return acc;
-          }, [])
-          .sort((a, b) => (a.title < b.title ? -1 : 1));
-
-  // const durations = [" ", "0:30",]
-  // for(let i=1; i<=12;i++){
-  //   durations.push(i+":00")
-  //   durations.push(i+":30")
-  // }
-
-  const durations = [" "];
-  const maxDurationHour = 48;
-  for (let i = 1; i < maxDurationHour; i++) {
-    durations.push(i + ":00");
-    // durations.push(i+":30")
-  }
-  // durations.pop()
-  durations.push(maxDurationHour + ":00 +");
-
-  const [message, setMessage] = useState({ type: "info", msg: "" });
-  const [open, setOpen] = useState(false);
 
   const handleClose = () => {
     setOpen(false);
@@ -133,159 +89,60 @@ export default function CarrierPage() {
     },
   };
 
-  const getCity = (lat, long) => {
-    const URL = `https://nominatim.openstreetmap.org/reverse?format=json&lat=${lat}&lon=${long}&accept-language=en`;
-    axios.get(URL).then((response) => console.log("getCity CarrierPAge", response.data));
+  const handleSubmitForm = (values) => {
+    submitForm(values);
   };
-
-  // !! Compute Distance Between current city and nearest City in autoComplete "From".
-  const getDefaultCity = () => {
-    cities_json.forEach((element) => {
-      // debugger;
-      let absLat = Math.round(latitude);
-      let elementAbsLat = Math.round(element.latitude);
-      let absLng = Math.round(longitude);
-      let elementAbsLng = Math.round(element.longitude);
-      if (absLat === elementAbsLat && absLng === elementAbsLng) {
-        console.log("est");
-      } else {
-        const getDistance = (cityLat, cityLng) => {
-          const R = 6371e3;
-          const φ1 = (latitude * Math.PI) / 180; // φ, λ in radians
-          const φ2 = (cityLat * Math.PI) / 180;
-          const Δφ = ((cityLat - latitude) * Math.PI) / 180;
-          const Δλ = ((cityLng - longitude) * Math.PI) / 180;
-          const a =
-            Math.sin(Δφ / 2) * Math.sin(Δφ / 2) +
-            Math.cos(φ1) * Math.cos(φ2) * Math.sin(Δλ / 2) * Math.sin(Δλ / 2);
-          const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-          const d = R * c; // in metres
-          return d;
-        };
-        const getNearestCity = (arr) => {
-          let lowest = {};
-          let temp;
-          arr.forEach((element) => {
-            if (Object.keys(lowest).length === 0 && lowest.constructor === Object) {
-              lowest = {
-                id: element.id,
-                Name: element.Name,
-                distanceBetween: element.distanceBetween,
-              };
-            } else {
-              temp = element.distanceBetween;
-              if (temp < lowest)
-                lowest = {
-                  id: element.id,
-                  Name: element.Name,
-                  distanceBetween: element.distanceBetween,
-                };
-            }
-          });
-          return lowest;
-        };
-        let intervals = cities_json.map((element) => {
-          return {
-            id: element.ID,
-            Name: element.name,
-            distanceBetween: getDistance(element.latitude, element.longitude),
-          };
-        });
-        //console.log(intervals);
-      }
-      // console.log(intervals);
-    });
-    // console.log("Nearest City", getNearestCity(intervals));
-  };
-
-  useEffect(() => {
-    let startPos;
-    const geoSuccess = function (position) {
-      startPos = position;
-      setLatitude(startPos.coords.latitude);
-      setLongitude(startPos.coords.longitude);
-      getCity(startPos.coords.latitude, startPos.coords.longitude);
-    };
-    // geolocation determination is temporary commented out. To turn it on, uncomment the line bellow
-    // navigator.geolocation.getCurrentPosition(geoSuccess);
-  }, []);
-
-  useEffect(() => {
-    if (latitude && longitude) {
-      getDefaultCity();
-    }
-  });
 
   return (
     // <Container maxWidth="sm" className={classes.drivePage}>
-    <Container maxWidth="sm" style={{ marginTop: "75px" }}>
+    <Container maxWidth='sm' style={{ marginTop: '75px' }}>
       <Formik
         initialValues={{
           date: new Date().toJSON().slice(0, 10),
-          departureTime: "",
+          departureTime: '',
           timeZone: userTimeZone.shift,
-          phoneNumber: "",
+          phoneNumber: '',
           places: 1,
-          price: "",
+          price: '',
           currency: rideCurrency,
-          duration: "",
+          duration: '',
           passAParcel: false,
           isPetsAllowed: false,
-          additionalInfo: "",
+          additionalInfo: '',
           regularTrips: false,
           regularTripsDays: {
             _0monday: {
               selected: false,
-              departureTime: "",
+              departureTime: '',
             },
             _1tuesday: {
               selected: false,
-              departureTime: "",
+              departureTime: '',
             },
             _2wednesday: {
               selected: false,
-              departureTime: "",
+              departureTime: '',
             },
             _3thursday: {
               selected: false,
-              departureTime: "",
+              departureTime: '',
             },
             _4friday: {
               selected: false,
-              departureTime: "",
+              departureTime: '',
             },
             _5saturday: {
               selected: false,
-              departureTime: "",
+              departureTime: '',
             },
             _6sunday: {
               selected: false,
-              departureTime: "",
+              departureTime: '',
             },
           },
         }}
         onSubmit={(values) => {
-          console.log("SUBMITTING");
-          const departureTimeGMT = values.departureTime.split(":");
-          departureTimeGMT[0] -= values.timeZone;
-          values.departureTime = departureTimeGMT.join(":");
-          console.log(values);
-          // dispatch(saveNewTransferAction(values));
-          // history.push("/");
-          dispatch(loadingUploadTransferAction(true));
-          uploadTransfer(values)
-            .then((response) => {
-              console.log(response);
-              setMessage({ type: "success", msg: i18n.t("SuccessTrip") });
-              dispatch(loadingUploadTransferAction(false));
-              setOpen(true);
-              // history.push("/");
-            })
-            .catch((error) => {
-              console.log(error);
-              setMessage({ type: "error", msg: error });
-              setOpen(true);
-            });
+          handleSubmitForm(values);
         }}
         validationSchema={schema}
       >
@@ -297,8 +154,10 @@ export default function CarrierPage() {
             schema.fields.duration = null;
             schema.fields.departureTime = null;
           } else {
-            schema.fields.date = yup.date().required("date.Required");
-            schema.fields.departureTime = yup.string().required("departureTime.Required");
+            schema.fields.date = yup.date().required('date.Required');
+            schema.fields.departureTime = yup
+              .string()
+              .required('departureTime.Required');
             // schema.fields.duration = yup.string().required("duration.Required");
           }
 
@@ -307,14 +166,15 @@ export default function CarrierPage() {
             // console.log(event.target.checked);
             const weekDays = {};
             Object.keys(props.values.regularTripsDays).map((weekDay) => {
-              return weekDays[weekDay] = {
+              return (weekDays[weekDay] = {
                 selected: event.target.checked,
-                departureTime: props.values.regularTripsDays[weekDay].departureTime,
-              };
+                departureTime:
+                  props.values.regularTripsDays[weekDay].departureTime,
+              });
             });
 
             // console.log("weekdays: ", weekDays);
-            props.setFieldValue("regularTripsDays", weekDays);
+            props.setFieldValue('regularTripsDays', weekDays);
           };
 
           return (
@@ -322,19 +182,19 @@ export default function CarrierPage() {
               <Autocomplete
                 className={styles.height95px}
                 {...defaultProps}
-                id="from"
-                name={"from"}
+                id='from'
+                name={'from'}
                 value={props.values.from}
-                margin="dense"
+                margin='dense'
                 onChange={(e, v) => {
-                  props.setFieldValue("from", v?.id || "");
+                  props.setFieldValue('from', v?.id || '');
                 }}
                 onBlur={props.handleBlur}
                 renderInput={(params) => (
                   <TextField
                     {...params}
-                    label={i18n.t("From")}
-                    margin="normal"
+                    label={i18n.t('From')}
+                    margin='normal'
                     error={Boolean(props.errors.from) && props.touched.from}
                     helperText={
                       Boolean(props.errors.from) &&
@@ -343,26 +203,28 @@ export default function CarrierPage() {
                     }
                   />
                 )}
-                ListboxProps={{ style: { maxHeight: "7rem" } }}
+                ListboxProps={{ style: { maxHeight: '7rem' } }}
               />
               <Autocomplete
                 className={styles.height95px}
                 {...defaultProps}
-                id="to"
-                name={"to"}
+                id='to'
+                name={'to'}
                 value={props.values.to}
                 onBlur={props.handleBlur}
                 onChange={(e, v) => {
-                  props.setFieldValue("to", v?.id || "");
+                  props.setFieldValue('to', v?.id || '');
                 }}
                 renderInput={(params) => (
                   <TextField
                     {...params}
-                    label={i18n.t("To")}
-                    margin="dense"
+                    label={i18n.t('To')}
+                    margin='dense'
                     error={Boolean(props.errors.to) && props.touched.to}
                     helperText={
-                      Boolean(props.errors.to) && props.touched.to && i18n.t(`form.errors.${props.errors.to}`)
+                      Boolean(props.errors.to) &&
+                      props.touched.to &&
+                      i18n.t(`form.errors.${props.errors.to}`)
                     }
                   />
                 )}
@@ -371,98 +233,127 @@ export default function CarrierPage() {
               <FormControlLabel
                 control={
                   <Checkbox
-                    id={"regularTrips"}
+                    id={'regularTrips'}
                     checked={props.values.regularTrips}
                     onChange={props.handleChange}
-                    color="primary"
+                    color='primary'
                   />
                 }
-                label={i18n.t("Regular trips")}
+                label={i18n.t('Regular trips')}
               />
               {props.values.regularTrips && (
-                <Paper variant="outlined" style={{ padding: "8px" }}>
+                <Paper variant='outlined' style={{ padding: '8px' }}>
                   <Grid
                     container
-                    direction="column"
+                    direction='column'
                     // alignItems="center"
                     // justify="center"
                     // style={{minHeight: "100vh"}}
                   >
-                    <Grid container direction={"row"}>
+                    <Grid container direction={'row'}>
                       <Grid item xs={9}>
                         <FormControlLabel
                           control={
                             <Checkbox
                               checked={
-                                Object.values(props.values.regularTripsDays).reduce(
+                                Object.values(
+                                  props.values.regularTripsDays
+                                ).reduce(
                                   (acc, val) => (acc += +val.selected),
                                   0
                                 ) === 7
                               }
                               onChange={handleSelectAllDaysChange}
-                              name="selectAll"
-                              margin={""}
+                              name='selectAll'
+                              margin={''}
                             />
                           }
-                          label={i18n.t("Select all")}
+                          label={i18n.t('Select all')}
                         />
                       </Grid>
                       <Grid item xs={3}>
-                        {i18n.t("Time")}
+                        {i18n.t('Time')}
                       </Grid>
                     </Grid>
 
-                    {Object.keys(props.values.regularTripsDays).map((weekDay) => {
-                      return (
-                        <Grid
-                          container
-                          direction={"row"}
-                          alignItems="flex-end"
-                          id={"regularTripsDays." + weekDay}
-                          key={"regularTripsDays." + weekDay}
-                        >
-                          <Grid item xs={9}>
-                            <FormControlLabel
-                              style={{ marginLeft: "10px" }}
-                              control={
-                                <Checkbox
-                                  id={"regularTripsDays." + weekDay + ".selected"}
-                                  checked={props.values.regularTripsDays[weekDay].selected}
-                                  onChange={props.handleChange}
-                                  name={"regularTripsDays." + weekDay + ".selected"}
-                                />
-                              }
-                              label={i18n.t(weekDay)}
-                            />
-                          </Grid>
-                          <Grid item xs={3}>
-                            <Tooltip title={i18n.t("Time")} placement="top">
-                              <TextField
-                                id={"regularTripsDays." + weekDay + ".departureTime"}
-                                name={"regularTripsDays." + weekDay + ".departureTime"}
-                                type="time"
-                                margin="normal"
-                                disabled={!props.values.regularTripsDays[weekDay].selected}
-                                value={props.values.regularTripsDays[weekDay].departureTime}
-                                onChange={props.handleChange}
+                    {Object.keys(props.values.regularTripsDays).map(
+                      (weekDay) => {
+                        return (
+                          <Grid
+                            container
+                            direction={'row'}
+                            alignItems='flex-end'
+                            id={'regularTripsDays.' + weekDay}
+                            key={'regularTripsDays.' + weekDay}
+                          >
+                            <Grid item xs={9}>
+                              <FormControlLabel
+                                style={{ marginLeft: '10px' }}
+                                control={
+                                  <Checkbox
+                                    id={
+                                      'regularTripsDays.' +
+                                      weekDay +
+                                      '.selected'
+                                    }
+                                    checked={
+                                      props.values.regularTripsDays[weekDay]
+                                        .selected
+                                    }
+                                    onChange={props.handleChange}
+                                    name={
+                                      'regularTripsDays.' +
+                                      weekDay +
+                                      '.selected'
+                                    }
+                                  />
+                                }
+                                label={i18n.t(weekDay)}
                               />
-                            </Tooltip>
+                            </Grid>
+                            <Grid item xs={3}>
+                              <Tooltip title={i18n.t('Time')} placement='top'>
+                                <TextField
+                                  id={
+                                    'regularTripsDays.' +
+                                    weekDay +
+                                    '.departureTime'
+                                  }
+                                  name={
+                                    'regularTripsDays.' +
+                                    weekDay +
+                                    '.departureTime'
+                                  }
+                                  type='time'
+                                  margin='normal'
+                                  disabled={
+                                    !props.values.regularTripsDays[weekDay]
+                                      .selected
+                                  }
+                                  value={
+                                    props.values.regularTripsDays[weekDay]
+                                      .departureTime
+                                  }
+                                  onChange={props.handleChange}
+                                />
+                              </Tooltip>
+                            </Grid>
                           </Grid>
-                        </Grid>
-                      );
-                    })}
+                        );
+                      }
+                    )}
                   </Grid>
                 </Paper>
               )}
               {!props.values.regularTrips && (
                 <>
-                  <Grid container justifyContent="space-between">
+                  <Grid container justifyContent='space-between'>
                     <Grid item xs={6}>
                       <TextField
-                        id="date"
-                        label={i18n.t("Date")}
-                        type="date"
-                        margin="normal"
+                        id='date'
+                        label={i18n.t('Date')}
+                        type='date'
+                        margin='normal'
                         fullWidth
                         onBlur={props.handleBlur}
                         error={Boolean(props.errors.date) && props.touched.date}
@@ -484,16 +375,19 @@ export default function CarrierPage() {
                     </Grid>
                   </Grid>
                   {/**** DEPARTURE TIME ****/}
-                  <Grid container justifyContent="space-between">
+                  <Grid container justifyContent='space-between'>
                     <Grid item xs={5}>
                       <TextField
-                        id="departureTime"
-                        label={i18n.t("Time")}
-                        type="time"
-                        margin="normal"
+                        id='departureTime'
+                        label={i18n.t('Time')}
+                        type='time'
+                        margin='normal'
                         fullWidth
                         onBlur={props.handleBlur}
-                        error={Boolean(props.errors.departureTime) && props.touched.departureTime}
+                        error={
+                          Boolean(props.errors.departureTime) &&
+                          props.touched.departureTime
+                        }
                         helperText={
                           Boolean(props.errors.departureTime) &&
                           props.touched.departureTime &&
@@ -509,72 +403,25 @@ export default function CarrierPage() {
                         }}
                       />
                     </Grid>
-                    {/*^^^ DEPARTURE TIME ^^^*/}
-
-                    {/*VVV DEPARTURE TIMEZONE   temporary commented out VVV*/}
-                    {/*<Grid item xs={6}>*/}
-                    {/*    <FormControl*/}
-                    {/*        fullWidth*/}
-                    {/*        style={{ paddingTop: "0px", marginTop: "9px" }}*/}
-                    {/*    >*/}
-                    {/*        <InputLabel*/}
-                    {/*            shrink*/}
-                    {/*            id="timeZone-label"*/}
-                    {/*            style={{ marginTop: "8px" }}*/}
-                    {/*        >*/}
-                    {/*            {i18n.t("Timezone")}*/}
-                    {/*        </InputLabel>*/}
-                    {/*        <Select*/}
-                    {/*            labelId="timeZone-label"*/}
-                    {/*            id="timeZone"*/}
-                    {/*            label="timeZone"*/}
-                    {/*            margin="normal"*/}
-                    {/*            name={"timeZone"}*/}
-                    {/*            value={userTimeZone.shift}*/}
-                    {/*            onChange={props.handleChange}*/}
-                    {/*            style={{ textTransform: "none" }}*/}
-                    {/*            helperText={" 123"}*/}
-                    {/*        >*/}
-                    {/*            {timeZones.map((item) => {*/}
-                    {/*                return (*/}
-                    {/*                    <MenuItem*/}
-                    {/*                        key={item.shift}*/}
-                    {/*                        value={item.shift}*/}
-                    {/*                        onClick={() => {*/}
-                    {/*                            setUserTimeZone(item);*/}
-                    {/*                        }}*/}
-                    {/*                    >*/}
-                    {/*                        /!*{item.name}*!/*/}
-                    {/*                        {"GMT+" +*/}
-                    {/*                            item.shift +*/}
-                    {/*                            " " +*/}
-                    {/*                            i18n.t("timezone." + item.shift)}*/}
-                    {/*                    </MenuItem>*/}
-                    {/*                );*/}
-                    {/*            })}*/}
-                    {/*        </Select>*/}
-                    {/*    </FormControl>*/}
-                    {/*</Grid>*/}
-                    {/*^^^ DEPARTURE TIMEZONE ^^^*/}
                   </Grid>
                 </>
               )}
-              <Grid container justifyContent="space-between">
+              <Grid container justifyContent='space-between'>
                 <Grid item xs={5}>
                   <FormControl fullWidth>
-                    <InputLabel shrink id="duration-label">
-                      {i18n.t("Travel time")}
+                    <InputLabel shrink id='duration-label'>
+                      {i18n.t('Travel time')}
                     </InputLabel>
                     <Select
-                      labelId="duration-label"
-                      id="duration"
-                      name={"duration"}
+                      labelId='duration-label'
+                      id='duration'
+                      name={'duration'}
                       value={props.values.duration}
                       renderValue={(value) => `${value}`}
-                      margin="dense"
+                      margin='dense'
                       // disableUnderline
                       onChange={props.handleChange}
-                      label="duration"
+                      label='duration'
                       //style={{paddingTop: "9px"}}
                     >
                       {durations.map((item) => {
@@ -593,19 +440,26 @@ export default function CarrierPage() {
                 </Grid>
               </Grid>
               {/* Phone number block */}
-              <Grid container justifyContent="space-between" alignItems="flex-end">
+              <Grid
+                container
+                justifyContent='space-between'
+                alignItems='flex-end'
+              >
                 {/* Phone */}
                 <Grid item xs={8}>
                   <TextField
                     fullWidth
-                    id="phoneNumber"
-                    name="phoneNumber"
-                    label={i18n.t("Phone number")}
-                    placeholder={"+1234567890"}
-                    margin="normal"
+                    id='phoneNumber'
+                    name='phoneNumber'
+                    label={i18n.t('Phone number')}
+                    placeholder={'+1234567890'}
+                    margin='normal'
                     value={props.values.phoneNumber}
                     onBlur={props.handleBlur}
-                    error={Boolean(props.errors.phoneNumber) && props.touched.phoneNumber}
+                    error={
+                      Boolean(props.errors.phoneNumber) &&
+                      props.touched.phoneNumber
+                    }
                     helperText={
                       Boolean(props.errors.phoneNumber) &&
                       props.touched.phoneNumber &&
@@ -614,76 +468,22 @@ export default function CarrierPage() {
                     onChange={props.handleChange}
                   />
                 </Grid>
-                {/* Messenger */}
-                <Grid item xs={4}>
-                  {/* Checkboxes */}
-                  {/* <FormGroup aria-label="position" row>
-                    <FormControlLabel
-                      control={
-                        <Checkbox
-                          checked={state.checkedB}
-                          // onChange={}
-                          name="checkedB"
-                          color="primary"
-                        />
-                      }
-                      labelPlacement="top"
-                      label="Telegram"
-                    />
-                    <FormControlLabel
-                      control={
-                        <Checkbox
-                          checked={state.checkedB}
-                          // onChange={}
-                          name="checkedB"
-                          color="primary"
-                        />
-                      }
-                      labelPlacement="top"
-                      label="WhatsApp"
-                    />
-                  </FormGroup> */}
-                  {/* Select */}
-                  {/* <InputLabel id="messengers">{i18n.t("Messenger")}</InputLabel>
-                  <Select
-                    className="select"
-                    id="messengers"
-                    labelId="messenger"
-                    name={"messengers"}
-                    margin="normal"
-                    disableUnderline
-                    onChange={({ target: { value } }) => {
-                      setMessenger(value);
-                    }}
-                  >
-                    <MenuItem value="Telegram">
-                      <img src={telegramIcon} alt="Telegram" />
-                    </MenuItem>
-                    <MenuItem value="WhatsApp">
-                      <img src={whatsAppIcon} alt="WhatsApp" />
-                    </MenuItem>
-                    <MenuItem value="Viber">
-                      <img src={viberIcon} alt="Viber" />
-                    </MenuItem>
-                  </Select> */}
-                  {/*  */}
-                </Grid>
               </Grid>
               {/*<Grid container >*/}
               <Grid
                 item
                 xs={12}
                 container
-                alignItems="flex-end"
-                style={{ display: "flex", alignItems: "center" }}
+                alignItems='flex-end'
+                style={{ display: 'flex', alignItems: 'center' }}
               >
                 <Grid item xs={8}>
                   <TextField
-                    type="number"
+                    type='number'
                     value={props.values.price}
-                    margin="dense"
-                    id="price"
-                    label={i18n.t("Price")}
+                    margin='dense'
+                    id='price'
+                    label={i18n.t('Price')}
                     onChange={props.handleChange}
                     onBlur={props.handleBlur}
                     error={Boolean(props.errors.price) && props.touched.price}
@@ -703,15 +503,15 @@ export default function CarrierPage() {
                 {/*CURRENCY v*/}
                 <Grid item xs={4}>
                   <Select
-                    id="currency"
-                    name={"currency"}
+                    id='currency'
+                    name={'currency'}
                     value={rideCurrency}
                     renderValue={(value) => `${value.toUpperCase()}`}
-                    margin="dense"
+                    margin='dense'
                     disableUnderline
                     onChange={props.handleChange}
-                    label="currency"
-                    style={{ paddingTop: "8px" }}
+                    label='currency'
+                    style={{ paddingTop: '8px' }}
                   >
                     {currencies.map((item) => {
                       return (
@@ -732,37 +532,40 @@ export default function CarrierPage() {
               <FormControlLabel
                 control={
                   <Checkbox
-                    id={"passAParcel"}
+                    id={'passAParcel'}
                     checked={props.values.passAParcel}
                     onChange={props.handleChange}
-                    color="primary"
+                    color='primary'
                     value={props.values.parcel}
                   />
                 }
-                label={i18n.t("Pass a parcel")}
+                label={i18n.t('Pass a parcel')}
               />
               <FormControlLabel
                 control={
                   <Checkbox
-                    id="isPetsAllowed"
+                    id='isPetsAllowed'
                     checked={props.values.isPetsAllowed}
                     onChange={props.handleChange}
-                    color="primary"
+                    color='primary'
                     value={props.values.isPetsAllowed}
                   />
                 }
-                label={i18n.t("PetsAllowed")}
+                label={i18n.t('PetsAllowed')}
               />
               <TextField
                 value={props.values.additionalInfo}
-                margin="normal"
-                id="additionalInfo"
-                name="additionalInfo"
+                margin='normal'
+                id='additionalInfo'
+                name='additionalInfo'
                 fullWidth
                 multiline
                 rows={2}
-                error={Boolean(props.errors.additionalInfo) && props.touched.additionalInfo}
-                label={i18n.t("Additional information")}
+                error={
+                  Boolean(props.errors.additionalInfo) &&
+                  props.touched.additionalInfo
+                }
+                label={i18n.t('Additional information')}
                 onChange={props.handleChange}
                 helperText={
                   Boolean(props.errors.additionalInfo) &&
@@ -770,11 +573,21 @@ export default function CarrierPage() {
                   i18n.t(`form.errors.${props.errors.additionalInfo}`)
                 }
               />
-              <div style={{ margin: "10px" }} className={"submitBtn"}>
-                <LoadingButton loading={loading} color="primary" variant="contained" fullWidth type="submit">
-                  {i18n.t("Publish a ride")}
+              <div style={{ margin: '10px' }} className={'submitBtn'}>
+                <LoadingButton
+                  loading={loading}
+                  color='primary'
+                  variant='contained'
+                  fullWidth
+                  type='submit'
+                >
+                  {i18n.t('Publish a ride')}
                 </LoadingButton>
-                <Dialog open={open} onClose={handleClose} aria-labelledby="responsive-dialog-title">
+                <Dialog
+                  open={open}
+                  onClose={handleClose}
+                  aria-labelledby='responsive-dialog-title'
+                >
                   <DialogContent>
                     <Alert severity={message.type}>{message.msg}</Alert>
                   </DialogContent>
