@@ -2,14 +2,19 @@ import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { FormControl, Select, MenuItem } from '@material-ui/core';
 import { currencies } from './currencies';
-import { getCurrency } from '../../../trip_search/presentation/redux/reducers/selectors';
-import { setCurrencyAction } from '../../../trip_search/presentation/redux/reducers/actions/app-actions';
-import {useStyles} from "../../../../general/MUI/useStyles";
+import { useStyles } from '../../../../general/MUI/useStyles';
+import {setCurrency} from "../../../../general/redux/slices/appSettingsSlice";
+import {getCurrency} from "../../../../general/redux/selectors";
 
 const CurrenciesSelector = () => {
   const classes = useStyles();
   const dispatch = useDispatch();
   const cur = useSelector(getCurrency);
+
+  const handleClick = (code) => {
+    localStorage.setItem('currency', code);
+    dispatch(setCurrency(code));
+  };
   return (
     <FormControl>
       <Select
@@ -20,18 +25,12 @@ const CurrenciesSelector = () => {
           },
         }}
         value={cur}
-        renderValue={value => `${value.toUpperCase()}`}
+        renderValue={(value) => `${value.toUpperCase()}`}
         disableUnderline
       >
-        {currencies.map(item => {
+        {currencies.map((item) => {
           return (
-            <MenuItem
-              key={item.code}
-              onClick={() => {
-                localStorage.setItem("currency", item.code);
-                dispatch(setCurrencyAction(item.code));
-              }}
-            >
+            <MenuItem key={item.code} onClick={() => handleClick(item.code)}>
               {item.code + `  ` + item.name}
             </MenuItem>
           );
